+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/theme/StreamDimens.kt
@@ -64,6 +64,8 @@ import androidx.compose.ui.unit.dp
  * @param quotedMessageAttachmentBottomPadding The bottom padding of the quoted message attachment preview.
  * @param quotedMessageAttachmentStartPadding The start padding of the quoted message attachment preview.
  * @param quotedMessageAttachmentEndPadding The end padding of the quoted message attachment preview.
+ * @param groupAvatarInitialsXOffset The x offset of the user initials inside avatar when there are more than two users.
+ * @param groupAvatarInitialsXOffset The y offset of the user initials inside avatar when there are more than two users.
  */
 @Immutable
 public data class StreamDimens(
@@ -108,6 +110,8 @@ public data class StreamDimens(
     public val quotedMessageAttachmentBottomPadding: Dp,
     public val quotedMessageAttachmentStartPadding: Dp,
     public val quotedMessageAttachmentEndPadding: Dp,
+    public val groupAvatarInitialsXOffset: Dp,
+    public val groupAvatarInitialsYOffset: Dp
 ) {
     public companion object {
         /**
@@ -157,6 +161,8 @@ public data class StreamDimens(
             quotedMessageAttachmentTopPadding = 6.dp,
             quotedMessageAttachmentStartPadding = 8.dp,
             quotedMessageAttachmentEndPadding = 0.dp,
+            groupAvatarInitialsXOffset = 1.5.dp,
+            groupAvatarInitialsYOffset = 2.5.dp
         )
     }
 }